v6.3.5

+++ Yodo1Suit/Purchase/Classes/Yodo1PurchaseDataAnalytics.m
@@ -56,7 +56,7 @@ static NSString* const __status                 = @"status";
     }
     [self.itemProperty setObject:product.uniformProductId ? :@"" forKey:__itemCode];
     [self.itemProperty setObject:product.productName ? :@"" forKey:__itemName];
-    [self.itemProperty setObject:[NSString stringWithFormat:@"%d",product.productType]  forKey:__itemType];
+    [self.itemProperty setObject:[NSString stringWithFormat:@"%ld",(long)product.productType]  forKey:__itemType];
     [self.itemProperty setObject:product.productPrice ? :@"" forKey:__itemPrice];
     [self.itemProperty setObject:product.currency ? :@"" forKey:__itemCurrency];
     [self.itemProperty setObject:product.channelProductId forKey:__channelItemCode];
